+++ packages/content-model/src/sveltia/index.ts
@@ -1,1 +1,2 @@
-export { createSveltiaCollection } from './adapter';
+export { createSveltiaCollection, createSveltiaCollections } from './adapter';
+export type { SveltiaCollectionOptions, SveltiaFieldCustomizationContext } from './adapter';